Spring2018-Project1-keran0107
Here I conduct a brief study about “What did the presidents say at their inauguation?”. In this project I explored the texts of U.S. presidents’ inaugrual speeches, from that of George Washington to that of Donald Trump which was delivered earlier this year. I focus on the differences between parties and terms to do basic analysis, topic model and sentient analysis.